refactor: [renderer] N8: move forceConstrainParent to plugins/d3Forces, delete dead SVG MicroForce runner
